Reformat the cog to original
