+++ koala/cogs/announce/cog.py
@@ -1,112 +1,117 @@
-# Own imports
+# Built-in/Generic Imports
 import math
 import time
 
+# Libs
 import discord
 from discord.ext import commands
 
+# Own modules
 import koalabot
-from koala.cogs.announce.utils import (ANNOUNCE_SEPARATION_DAYS,
-                                       MAX_MESSAGE_LENGTH, SECONDS_IN_A_DAY)
-from koala.log import logger
+from koala.colours import KOALA_GREEN
+from koala.db import insert_extension
 from koala.utils import extract_id, wait_for_message
 
-from .announce_message import Announce, AnnounceMessage
+from .announce_message import AnnounceMessage
+from .db import AnnounceDBManager
+from .log import logger
+from .utils import (ANNOUNCE_SEPARATION_DAYS, MAX_MESSAGE_LENGTH,
+                    SECONDS_IN_A_DAY)
 
 
-def announce_is_enabled(guild):
-    return announce_is_enabled(guild)
-
-
-def not_exceeded_limit(self, guild_id, ctx):
+def announce_is_enabled(ctx):
     """
-    Check if the number of announcements in the guild is not exceeded
-    :param guild_id: The id of the guild
+    A command used to check if the guild has enabled announce
+    e.g. @commands.check(announce_is_enabled)
+
     :param ctx: The context of the message
-    :return: True if the number of announcements is not exceeded, False otherwise
+    :return: True if enabled or test, False otherwise
     """
     try:
-        result = self.Announce.not_exceeded_limit(guild_id, ctx)
+        result = koalabot.check_guild_has_ext(ctx, "Announce")
     except PermissionError:
         result = False
 
-    ctx.send(result or (str(ctx.guild) == koalabot.TEST_USER and koalabot.is_dpytest))
+    return result or (str(ctx.guild) == koalabot.TEST_USER and koalabot.is_dpytest)
 
 
-def has_active_msg(self, guild_id, ctx):
+class Announce(commands.Cog):
     """
-    Check if there is an active announcement message
-    :param guild_id: The id of the guild
-    :param ctx: The context of the message
-    :return: True if there is an active announcement message, False otherwise
+        Send DM announcements to certain roles and people.
     """
-    try:
-        result = self.Announce.has_active_msg(guild_id, ctx)
-    except PermissionError:
-        result = False
 
-    ctx.send(result or (str(ctx.guild) == koalabot.TEST_USER and koalabot.is_dpytest))
+    def __init__(self, bot):
+        self.bot = bot
+        self.messages = {}
+        self.roles = {}
+        insert_extension("Announce", 0, True, True)
+        self.announce_database_manager = AnnounceDBManager()
 
+    def not_exceeded_limit(self, guild_id):
+        """
+        Check if enough days have passed for the user to use the announce function
+        :return:
+        """
+        if self.announce_database_manager.get_last_use_date(guild_id):
+            return int(time.time()) - self.announce_database_manager.get_last_use_date(
+                guild_id) > ANNOUNCE_SEPARATION_DAYS * SECONDS_IN_A_DAY
+        return True
 
-def get_role_names(self, guild_id, roles, ctx):
-    """
-    Get the names of the roles
-    :param guild_id: The id of the guild
-    :param roles: The roles
-    :param ctx: The context of the message
-    :return: A string consisting the names of the roles
-    """
-    try:
-        result = self.Announce.get_role_names(guild_id, roles, ctx)
-    except PermissionError:
-        result = False
+    def has_active_msg(self, guild_id):
+        """
+        Check if a particular id has an active announcement pending announcement
+        :param guild_id: The id of the guild of the command
+        :return: Boolean of whether there is an active announcement or not
+        """
+        return guild_id in self.messages.keys()
 
-    ctx.send(result or (str(ctx.guild) == koalabot.TEST_USER and koalabot.is_dpytest))
+    def get_role_names(self, guild_id, roles):
+        """
+        A function to get the names of all the roles the announcement will be sent to
+        :param roles: The list of roles in the guild
+        :param guild_id: The id of the guild
+        :return: All the names of the roles that are tagged
+        """
+        temp = []
+        for role in self.roles[guild_id]:
+            temp.append(discord.utils.get(roles, id=role).name)
+        return temp
 
+    def get_receivers(self, guild_id, roles):
+        """
+        A function to get the receivers of a particular announcement
+        :param roles: The list of roles in the guild
+        :param guild_id: The id of the guild
+        :return: All the receivers of the announcement
+        """
+        temp = []
+        for role in self.roles[guild_id]:
+            temp += discord.utils.get(roles, id=role).members
+        return list(set(temp))
 
-def get_receivers(self, guild_id, roles, ctx):
-    """
-    Get the receivers of the announcement
-    :param guild_id: The id of the guild
-    :param roles: The roles
-    :param ctx: The context of the message
-    :return: A list of the receivers
-    """
-    try:
-        result = self.Announce.get_receivers(guild_id, roles, ctx)
-    except PermissionError:
-        result = False
-
-    ctx.send(result or (str(ctx.guild) == koalabot.TEST_USER and koalabot.is_dpytest))
-
-
-def receiver_msg(self, guild, ctx):
-    """
+    def receiver_msg(self, guild):
+        """
         A function to create a string message about receivers
         :param guild: The guild of the bot
         :return: A string message about receivers
         """
-    try:
-        result = self.Announce.receiver_msg(guild, ctx)
-    except PermissionError:
-        result = False
+        if not self.roles[guild.id]:
+            return f"You are currently sending to Everyone and there are {str(len(guild.members))} receivers"
+        return f"You are currently sending to {self.get_role_names(guild.id, guild.roles)} and there are {str(len(self.get_receivers(guild.id, guild.roles)))} receivers "
 
-    ctx.send(result or (str(ctx.guild) == koalabot.TEST_USER and koalabot.is_dpytest))
-
-
-def construct_embed(self, guild: discord.Guild, ctx):
-    """
-    Construct an embed message
-    :param guild: The guild of the bot
-    :param ctx: The context of the message
-    :return: An embed message
-    """
-    try:
-        result = self.Announce.construct_embed(guild, ctx)
-    except PermissionError:
-        result = False
-
-    ctx.send(result or (str(ctx.guild) == koalabot.TEST_USER and koalabot.is_dpytest))
+    def construct_embed(self, guild: discord.Guild):
+        """
+        Constructing an embedded message from the information stored in the manager
+        :param guild: The the guild
+        :return: An embedded message for the announcement
+        """
+        message = self.messages[guild.id]
+        embed: discord.Embed = discord.Embed(title=message.title,
+                                             description=message.description, colour=KOALA_GREEN)
+        embed.set_author(name="Announcement from " + guild.name)
+        if message.thumbnail != 'https://cdn.discordapp.com/':
+            embed.set_thumbnail(url=message.thumbnail)
+        return embed
 
     @commands.check(announce_is_enabled)
     @commands.group(name="announce")
@@ -202,8 +207,7 @@ def construct_embed(self, guild: discord.Guild, ctx):
         :return:
         """
         if self.has_active_msg(ctx.guild.id):
-            await ctx.send(
-                "Please enter the roles you want to tag separated by space, I'll wait for 60 seconds, no rush.")
+            await ctx.send("Please enter the roles you want to tag separated by space, I'll wait for 60 seconds, no rush.")
             message, channel = await wait_for_message(self.bot, ctx)
             if not message:
                 await channel.send("Okay, I'll cancel the command.")
@@ -226,8 +230,7 @@ def construct_embed(self, guild: discord.Guild, ctx):
         :return:
         """
         if self.has_active_msg(ctx.guild.id):
-            await ctx.send(
-                "Please enter the roles you want to remove separated by space, I'll wait for 60 seconds, no rush.")
+            await ctx.send("Please enter the roles you want to remove separated by space, I'll wait for 60 seconds, no rush.")
             message, channel = await wait_for_message(self.bot, ctx)
             if not message:
                 await channel.send("Okay, I'll cancel the command.")
@@ -269,13 +272,13 @@ def construct_embed(self, guild: discord.Guild, ctx):
                     try:
                         await receiver.send(embed=embed)
                     except (discord.Forbidden, AttributeError, discord.HTTPException) as e:
-                        logger.error(f'User {receiver.id} cannot receive dms')
+                        logger.error(f'User {receiver.id} cannot recieve dms')
             else:
                 for receiver in ctx.guild.members:
                     try:
                         await receiver.send(embed=embed)
                     except (discord.Forbidden, AttributeError, discord.HTTPException) as e:
-                        logger.error(f'User {receiver.id} cannot receive dms')
+                        logger.error(f'User {receiver.id} cannot recieve dms')
 
             self.messages.pop(ctx.guild.id)
             self.roles.pop(ctx.guild.id)
